Bug fix
Fixed a class cast exception that was caused by the user details implementation class, because it wouldn't be cast to my custom users object; even though they both implement the same class. The problem was solved by making the load by user name method in the user details service return a value of type user; because the object can only be cast successfully if they have an inheritance relationship.